job array

+++ render-blender.py
@@ -2,7 +2,6 @@ import argparse, sys, os, math, re
 import bpy
 from glob import glob
 import random
-import multiprocessing
 
 ## blender --background --python render_blender.py -- --output_folder /tmp path_to_model.obj ##
 #test
@@ -34,13 +33,16 @@ parser.add_argument('--random', type=bool, default=False,
                     help='Randomize selected objects')
 parser.add_argument('--num', type=int, default=100,
                     help='number of object to render, to use only with random')
-parser.add_argument('--node', type=int, default=1,
-                    help='define the number of nodes used to render the objects')
+parser.add_argument('--job_id', type=int, default=1,
+                    help='gives job id')
+parser.add_argument('--num_job', type=int, default=10,
+                    help='number of jobs')
 
 
 argv = sys.argv[sys.argv.index("--") + 1:]
 args = parser.parse_args(argv)
 
+# <editor-fold desc="Description">
 # Set up rendering
 context = bpy.context
 scene = bpy.context.scene
@@ -184,6 +186,7 @@ except:
 
 # Import textured mesh
 bpy.ops.object.select_all(action='DESELECT')
+# </editor-fold>
 
 directory = args.obj
 j = 0
@@ -191,9 +194,17 @@ j = 0
 
 list = glob(directory + '/**/*.obj', recursive=True)
 
-def render():
-        name = random.choice(list)
+if args.random == True:
+    bpy.context.scene.render.use_persistent_data = False
+    N = args.num
+    New_list = []
+    for pp in range(N):
+        New_list.append(randome.choice(list))
+    part = N//args.num_job
 
+
+    for p in range((args.job_id-1)*part, (args.job_id)*part):
+        name = New_list[p]
         w = 0
         bpy.ops.object.select_by_type(type='MESH')
         bpy.ops.object.delete()
@@ -290,21 +301,6 @@ def render():
             bpy.ops.render.render(write_still=True)  # render still
             bpy.context.scene.render.use_persistent_data = True
             cam_empty.rotation_euler[2] += math.radians(stepsize)
-
-
-
-if args.random == True:
-    bpy.context.scene.render.use_persistent_data = False
-    N = args.num
-    for p in range(N):
-        if args.node==1:
-            render()
-        else:
-            jobs=[]
-            for i in range(args.node):
-                p=multiprocessing.Process(target=render)
-                jobs.append(p)
-                p.start()
 
 else:
     for name in glob(directory + '/**/*.obj', recursive=True):
